Further cleanup of A_InsertSprite().
git-svn-id: https://svn.eduke32.com/eduke32@5567 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/game.h
@@ -318,7 +318,7 @@ extern uint32_t g_frameDelay;
 extern user_defs ud;
 
 int32_t A_CheckInventorySprite(spritetype *s);
-int32_t A_InsertSprite(int32_t whatsect,int32_t s_x,int32_t s_y,int32_t s_z,int32_t s_pn,int32_t s_s,int32_t s_xr,int32_t s_yr,int32_t s_a,int32_t s_ve,int32_t s_zv,int32_t s_ow,int32_t s_ss);
+int32_t A_InsertSprite(int16_t whatsect,int32_t s_x,int32_t s_y,int32_t s_z,int16_t s_pn,int8_t s_s,uint8_t s_xr,uint8_t s_yr,int16_t s_a,int16_t s_ve,int16_t s_zv,int16_t s_ow,int16_t s_ss);
 int32_t A_Spawn(int32_t j,int32_t pn);
 int32_t G_DoMoveThings(void);
 //int32_t G_EndOfLevel(void);
